+++ main/plugins/org.talend.designer.core.generic/src/main/java/org/talend/designer/core/generic/service/GenericService.java
@@ -22,6 +22,7 @@ import org.talend.commons.runtime.model.components.IComponentConstants;
 import org.talend.components.api.component.ComponentDefinition;
 import org.talend.components.api.properties.ComponentProperties;
 import org.talend.components.api.properties.ComponentReferenceProperties;
+import org.talend.components.api.service.ComponentService;
 import org.talend.components.api.wizard.ComponentWizardDefinition;
 import org.talend.core.GlobalServiceRegister;
 import org.talend.core.model.components.IComponent;
@@ -30,12 +31,18 @@ import org.talend.core.model.metadata.builder.connection.MetadataTable;
 import org.talend.core.model.process.EParameterFieldType;
 import org.talend.core.model.process.IElementParameter;
 import org.talend.core.model.process.INode;
+import org.talend.core.model.repository.ERepositoryObjectType;
+import org.talend.core.runtime.services.IGenericDBService;
 import org.talend.core.runtime.services.IGenericService;
+import org.talend.core.runtime.services.IGenericWizardInternalService;
 import org.talend.core.runtime.services.IGenericWizardService;
+import org.talend.core.utils.ReflectionUtils;
 import org.talend.daikon.properties.Properties;
 import org.talend.designer.core.generic.model.Component;
 import org.talend.designer.core.generic.model.GenericElementParameter;
 import org.talend.designer.core.generic.utils.ComponentsUtils;
+import org.talend.repository.model.IRepositoryNode.ENodeType;
+import org.talend.repository.model.RepositoryNode;
 
 import orgomg.cwm.objectmodel.core.TaggedValue;
 
@@ -57,6 +64,55 @@ public class GenericService implements IGenericService {
             }
         }
         return typeNames;
+    }
+
+    @Override
+    public List<RepositoryNode> createNodesFromComponentService(RepositoryNode curParentNode) {
+        List<RepositoryNode> repNodes = new ArrayList<>();
+        ComponentService componentService = ComponentsUtils.getComponentService();
+        Set<ComponentWizardDefinition> wizardDefinitions = componentService.getTopLevelComponentWizards();
+        IGenericWizardInternalService interService = IGenericWizardInternalService.getService();
+        for (ComponentWizardDefinition wizardDefinition : wizardDefinitions) {
+            String name = wizardDefinition.getName();
+            String displayName = wizardDefinition.getDisplayName();
+            String folder = "metadata/" + name; //$NON-NLS-1$
+            int ordinal = 100;
+            ERepositoryObjectType repositoryType = interService.createRepositoryType(name, displayName, name, folder, ordinal);
+            if (curParentNode == null && "JDBC".equals(name)) { //$NON-NLS-1$
+                Class<ComponentProperties> jdbcClass = ReflectionUtils.getClass(
+                        "org.talend.components.jdbc.wizard.JDBCConnectionWizardProperties",
+                        wizardDefinition.getClass().getClassLoader());
+                if (jdbcClass != null && wizardDefinition.supportsProperties(jdbcClass)) {
+                    IGenericDBService dbService = null;
+                    if (GlobalServiceRegister.getDefault().isServiceRegistered(IGenericDBService.class)) {
+                        dbService = (IGenericDBService) GlobalServiceRegister.getDefault().getService(IGenericDBService.class);
+                    }
+                    if (dbService != null) {
+                        dbService.getExtraTypes().add(repositoryType);
+                    }
+                }
+            }
+            if (curParentNode != null && !needHide(repositoryType)) {
+                repNodes.add(interService.createRepositoryNode(curParentNode, wizardDefinition.getDisplayName(), repositoryType,
+                        ENodeType.SYSTEM_FOLDER));
+            }
+        }
+        return repNodes;
+    }
+
+    private boolean needHide(ERepositoryObjectType type) {
+        if (type == null) {
+            return false;
+        }
+        List<ERepositoryObjectType> extraTypes = new ArrayList<ERepositoryObjectType>();
+        IGenericDBService dbService = null;
+        if (GlobalServiceRegister.getDefault().isServiceRegistered(IGenericDBService.class)) {
+            dbService = (IGenericDBService) GlobalServiceRegister.getDefault().getService(IGenericDBService.class);
+        }
+        if (dbService != null) {
+            extraTypes.addAll(dbService.getExtraTypes());
+        }
+        return extraTypes.contains(type);
     }
 
     @Override
