+++ main/plugins/org.talend.repository.generic/src/main/java/org/talend/repository/generic/internal/service/GenericWizardInternalService.java
@@ -20,8 +20,8 @@ import org.talend.components.api.properties.ComponentProperties;
 import org.talend.components.api.service.ComponentService;
 import org.talend.components.api.wizard.ComponentWizard;
 import org.talend.core.model.repository.ERepositoryObjectType;
+import org.talend.core.runtime.services.IGenericWizardInternalService;
 import org.talend.designer.core.generic.utils.ComponentsUtils;
-import org.talend.repository.generic.internal.IGenericWizardInternalService;
 import org.talend.repository.model.IRepositoryNode.ENodeType;
 import org.talend.repository.model.IRepositoryNode.EProperties;
 import org.talend.repository.model.RepositoryNode;
