+++ main/plugins/org.talend.repository.generic/src/main/java/org/talend/repository/generic/model/FakeGenericExtendNode.java
@@ -2,8 +2,7 @@ package org.talend.repository.generic.model;
 
 import org.talend.commons.ui.runtime.image.IImage;
 import org.talend.commons.ui.runtime.repository.IExtendRepositoryNode;
-import org.talend.core.GlobalServiceRegister;
-import org.talend.core.runtime.services.IGenericWizardService;
+import org.talend.core.runtime.services.IGenericService;
 
 /**
  * 
@@ -17,10 +16,7 @@ public class FakeGenericExtendNode implements IExtendRepositoryNode {
     }
 
     private void initGenericTypes() {
-        IGenericWizardService wizardService = null;
-        if (GlobalServiceRegister.getDefault().isServiceRegistered(IGenericWizardService.class)) {
-            wizardService = (IGenericWizardService) GlobalServiceRegister.getDefault().getService(IGenericWizardService.class);
-        }
+        IGenericService wizardService = IGenericService.getService();
         if (wizardService != null) {
             wizardService.createNodesFromComponentService(null); // Just use to init generic types.
         }
